restructuring front end

+++ devices/airspy-handler/airspy-handler.cpp
@@ -136,7 +136,7 @@ uint32_t samplerate_count;
 	my_airspy_get_samplerates (device, myBuffer, samplerate_count);
 
 	selectedRate	= 0;
-	for (i = 0; i < samplerate_count; i ++) {
+	for (i = 0; i <  (int)samplerate_count; i ++) {
 	   fprintf (stderr, "%d \n", myBuffer [i]);
 	   if (abs (myBuffer [i] - 2048000.0) < distance) {
 	      distance	= abs (myBuffer [i] - 2048000.0);
